dom/network/tests/test_tcpsocket_default_permissions.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test to ensure TCPSocket permission is disabled by default</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script type="application/javascript">

/** Test to ensure TCPSocket permission is disabled by default **/

var caught = false;
try {
  new TCPSocket("localhost", 80, {})
} catch (e) {
  caught = true;
}

ok(caught, "TCPSocket should not exist by default");

var caught = false;
try {
  navigator.mozTCPSocket.open("localhost", 80, {})
} catch (e) {
  caught = true;
}

ok(caught, "navigator.mozTCPSocket.open should not exist by default");

</script>
</pre>
</body>
</html>
